// galaauction.client/src/types/Guest.tsx
export type GuestType = {
    guestId: number;
    firstName: string;
    lastName: string;
    tableNumber: number;
    galaEventId: number;
    inPersonBidderNumber: number;
    onlineBidderNumber: number;
    onlineBidderOnly: boolean;
    inPersonAutoGen: boolean;
    onlineAutoGen: boolean;
    fullName: string;
    fullNameReversed: string;
};